+++ src/viewerRegistry.ts
@@ -36,6 +36,7 @@ import { safetensorsViewer } from './viewers/safetensorsViewer';
 import { ggufViewer } from './viewers/ggufViewer';
 import { onnxViewer } from './viewers/onnxViewer';
 import { tfliteViewer } from './viewers/tfliteViewer';
+import { kerasViewer } from './viewers/kerasViewer';
 import { latexViewer } from './viewers/latexViewer';
 
 export const VIEWER_DEFINITIONS: ViewerDefinition[] = [
@@ -74,6 +75,7 @@ export const VIEWER_DEFINITIONS: ViewerDefinition[] = [
     ggufViewer,
     onnxViewer,
     tfliteViewer,
+    kerasViewer,
     hwpViewer,
     psdViewer,
     excelViewer,